pyro_env_requirements.txt
# packages in environment at /home/fcairoli/miniconda3/envs/activeclone:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
backports.weakref         1.0rc1                   py36_0    marta-sd
blas                      1.0                         mkl  
ca-certificates           2019.11.27                    0  
cairo                     1.14.12              h8948797_3  
certifi                   2019.11.28               py36_0  
contextlib2               0.6.0.post1                py_0  
cudatoolkit               10.1.243             h6bb024c_0  
cudnn                     5.1                           0    marta-sd
cycler                    0.10.0                   py36_0  
dbus                      1.13.12              h746ee38_0  
edward                    1.3.5                    pypi_0    pypi
expat                     2.2.6                he6710b0_0  
fontconfig                2.13.0               h9420a91_0  
freetype                  2.9.1                h8a8886c_1  
fribidi                   1.0.5                h7b6447c_0  
glib                      2.63.1               h5a9c865_0  
graphite2                 1.3.13               h23475e2_0  
graphviz                  2.40.1               h21bd128_2  
gst-plugins-base          1.14.0               hbbd80ab_1  
gstreamer                 1.14.0               hb453b48_1  
h5py                      2.9.0            py36h7918eee_0  
harfbuzz                  1.8.8                hffaf4a1_0  
hdf5                      1.10.4               hb1b8bf9_0  
icu                       58.2                 h9c2bf20_1  
intel-openmp              2019.4                      243  
joblib                    0.14.1                     py_0  
jpeg                      9b                   h024ee3a_2  
keras                     2.2.4                         0  
keras-applications        1.0.8                      py_0  
keras-base                2.2.4                    py36_0  
keras-preprocessing       1.1.0                      py_1  
kiwisolver                1.1.0            py36he6710b0_0  
libedit                   3.1.20181209         hc058e9b_0  
libffi                    3.2.1                hd88cf55_4  
libgcc-ng                 9.1.0                hdf63c60_0  
libgfortran-ng            7.3.0                hdf63c60_0  
libpng                    1.6.37               hbc83047_0  
libstdcxx-ng              9.1.0                hdf63c60_0  
libtiff                   4.1.0                h2733197_0  
libuuid                   1.0.3                h1bed415_2  
libxcb                    1.13                 h1bed415_1  
libxml2                   2.9.9                hea5a465_1  
markdown                  3.1.1                    py36_0  
matlabengineforpython     R2018a                   pypi_0    pypi
matplotlib                3.1.1            py36h5429711_0  
mkl                       2019.4                      243  
mkl-service               2.3.0            py36he904b0f_0  
mkl_fft                   1.0.15           py36ha843d7b_0  
mkl_random                1.1.0            py36hd6b4f25_0  
mock                      3.0.5                    py36_0  
ncurses                   6.1                  he6710b0_1  
ninja                     1.9.0            py36hfd86e86_0  
numpy                     1.17.4                   pypi_0    pypi
numpy-base                1.17.4           py36hde5b4d6_0  
olefile                   0.46                     py36_0  
openssl                   1.1.1d               h7b6447c_3  
opt_einsum                3.1.0                      py_0  
pandas                    0.25.3           py36he6710b0_0  
pango                     1.42.4               h049681c_0  
patsy                     0.5.1                    py36_0  
pcre                      8.43                 he6710b0_0  
pillow                    7.0.0            py36hb39fc2d_0  
pip                       19.3.1                   py36_0  
pixman                    0.38.0               h7b6447c_0  
protobuf                  3.2.0                    py36_0    marta-sd
pyparsing                 2.4.5                      py_0  
pyqt                      5.9.2            py36h05f1152_2  
pyro-ppl                  0.3.4                    py36_0    tallic
python                    3.6.9                h265db76_0  
python-dateutil           2.8.1                      py_0  
python-graphviz           0.13.2                     py_0  
pytorch                   1.4.0           py3.6_cuda10.1.243_cudnn7.6.3_0    pytorch
pytz                      2019.3                     py_0  
pyyaml                    5.2              py36h7b6447c_0  
qt                        5.9.7                h5867ecd_1  
readline                  7.0                  h7b6447c_5  
scikit-learn              0.22             py36hd81dba3_0  
scipy                     1.3.2            py36h7c811a0_0  
setuptools                42.0.2                   py36_0  
sip                       4.19.8           py36hf484d3e_0  
six                       1.13.0                   py36_0  
sqlite                    3.30.1               h7b6447c_0  
statsmodels               0.10.1           py36hdd07704_0  
tensorflow                1.2.0            py36hb9c984a_3    marta-sd
tk                        8.6.8                hbc83047_0  
torchvision               0.5.0                py36_cu101    pytorch
tornado                   6.0.3            py36h7b6447c_0  
tqdm                      4.40.2                     py_0  
werkzeug                  0.16.0                     py_0  
wheel                     0.33.6                   py36_0  
xz                        5.2.4                h14c3975_4  
yaml                      0.1.7                had09818_2  
zlib                      1.2.11               h7b6447c_3  
zstd                      1.3.7                h0b5b093_0  
